// src/app/pages/operaciones/training/Assigned-course-registration-interface.ts
export interface AssignedCourseRegistration {
  idAssignedCourseRegistration: number;
  idRegisterFollow: number;
  advancePercentage: number;
  durationCourseHrs: number;
  durationCourseMin: number;
  nameCourse: string;
  observation: string;
  startDate: string;
  urlCourse: string;
  idAssignedCourseRegistrationDetail: number;
}